Add files via upload
After you buy it he send a file in .exe that is actually python but compiled, you can easily decrypt it with decompyler and you find out that is just a basic ui no codes just a basic ui after that he says its doesnt work becouse of the special proxies you have to use and YES you guessed it right you have to pay for it. But you can see the whole conversation.